requirements.txt
altgraph==0.10.2
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bdist-mpkg==0.5.0
bonjour-py==0.3
certifi==2020.4.5.1
cffi==1.12.2
chardet==3.0.4
click==7.1.2
configparser==3.5.0
cryptography==2.6.1
enum34==1.1.6
Flask==1.1.2
future==0.17.1
futures==3.2.0
idna==2.9
ipaddress==1.0.22
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
lazy-object-proxy==1.3.1
macholib==1.5.1
MarkupSafe==1.1.1
matplotlib==1.3.1
mccabe==0.6.1
modulegraph==0.10.4
nltk==3.5
numpy==1.8.0rc1
Pillow==6.2.2
protobuf==3.12.2
py2app==0.7.3
pycparser==2.19
PyJWT==1.7.1
pylint==1.9.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
regex==2020.4.4
requests==2.23.0
scipy==0.13.0b1
singledispatch==3.4.0.3
six==1.11.0
tflearn==0.3.2
tqdm==4.46.1
twilio==6.41.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.10.11
xattr==0.6.4

python-dotenv~=0.13.0